+++ lib/src/persistence_base.dart
@@ -1,6 +1,23 @@
+import 'package:simple_json_persistence/src/persistence_noop.dart'
+    if (dart.library.io) 'package:simple_json_persistence/src/persistence_io.dart'
+    if (dart.library.html) 'package:simple_json_persistence/src/persistence_html.dart';
+
 typedef BaseDirectoryBuilder = Future<String> Function();
 
 abstract class StoreBackend {
+  /// Allows overwriting the default mechanism for fetching the base
+  /// directory for storage (When not explicitly passed into
+  /// [create])
+  static BaseDirectoryBuilder defaultBaseDirectoryBuilder;
+
+  /// Creates a new store backend instance depending on either
+  /// dart:io or dart:html.
+  /// With dart:io the given [baseDirectoryBuilder] is used to determine
+  /// the file system location store json files. Otherwise
+  /// [defaultBaseDirectoryBuilder] is used.
+  static StoreBackend create([BaseDirectoryBuilder baseDirectoryBuilder]) =>
+      createStoreBackend(baseDirectoryBuilder);
+
   Future<Store> storeForFile(String name);
 }